+++ lib/exports.dart
@@ -9,3 +9,4 @@ export 'package:bbox_editor/src/bbox_fit_cover_mapper.dart';
 export 'package:bbox_editor/src/bbox_editor_events.dart';
 export 'package:bbox_editor/src/bbox_helpers.dart';
 export 'package:bbox_editor/src/bbox_overlay.dart';
+export 'package:bbox_editor/src/polygon_entity.dart';