+++ lib/src/bbox_editor_events.dart
@@ -25,4 +25,30 @@ class BoxSelected extends BBoxEvent {
   const BoxSelected({this.box, required super.origin});
 }
 
-class BoxesCleared extends BBoxEvent { const BoxesCleared({required super.origin}); }
+class BoxesCleared extends BBoxEvent {
+  const BoxesCleared({required super.origin});
+}
+
+sealed class PolygonEvent {
+  final CommitOrigin origin;
+  const PolygonEvent({required this.origin});
+}
+
+class PolygonCreated extends PolygonEvent {
+  final PolygonEntity polygon;
+  const PolygonCreated({required this.polygon, required super.origin});
+}
+
+class PolygonUpdated extends PolygonEvent {
+  final PolygonEntity polygon;
+  const PolygonUpdated({required this.polygon, required super.origin});
+}
+
+class PolygonDeleted extends PolygonEvent {
+  final int id;
+  const PolygonDeleted({required this.id, required super.origin});
+}
+
+class PolygonsCleared extends PolygonEvent {
+  const PolygonsCleared({required super.origin});
+}